+++ sdk/communication/Azure.Communication.PhoneNumbers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 1.5.0-beta1 (Unreleased)
+
+### Features Added
+- Adds support for mobile number types
+  - mobile numbers are location associated phone numbers with SMS capabilities
+- API version `2025-06-01` is the default.
+
 ## 1.4.0 (2025-06-20)
 
 ### Features Added